2.21 KR (Please read description if necessary)
@ingbrzy @MarkH-UK @translators.xiaomi.eu

This part is done by our team as a localization process for Korean users, and our translator 'cjhyuky' told me that this lines were noticed to ingbrzy and you agreed. Please do not remove this strings since the feedback shows that a lot of Korean users are actually using this feature on Mi Browser.

And as always, Gamsa-hapnida! (meaning Thank you in Korean 👍 )

-XiaomiQM